move cache serialization tests to kernel tests
